Add a compatibility check for buildifier.
